--- src/resources/users/user.memory.repository.js
@@ -1,6 +1,0 @@
-const getAll = async () => {
-  // TODO: mock implementation. should be replaced during task development
-  return [];
-};
-
-module.exports = { getAll };
