--- src/resources/users/user.router.js
@@ -1,11 +1,0 @@
-const router = require('express').Router();
-const User = require('./user.model');
-const usersService = require('./user.service');
-
-router.route('/').get(async (req, res) => {
-  const users = await usersService.getAll();
-  // map user fields to exclude secret fields like "password"
-  res.json(users.map(User.toResponse));
-});
-
-module.exports = router;
